backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
openai>=1.3.0
python-dotenv==1.0.0
google-genai>=1.0.0
Pillow==10.2.0
python-multipart==0.0.9
boto3==1.34.0
sqlalchemy==2.0.25
asyncpg==0.29.0
arq==0.25.0
redis==5.0.1
slowapi==0.1.9
python-jose[cryptography]==3.3.0
httpx==0.27.0
sentry-sdk>=1.40.0
python-json-logger>=2.0.7
